+++ classes/Router.php
@@ -170,11 +170,23 @@ class Router
      * Add a controller route
      *
      * @param string $url
+     * @param string $handler
+     */
+    public function any($url, $handler)
+    {
+        $this->add('ANY', $this->split($url), $handler);
+    }
+
+    /**
+     * Add a controller route
+     *
+     * @param string $url
      * @param string $class
+     * @deprecated Use any($url, $handler) instead.
      */
     public function controller($url, $class)
     {
-        $this->add('ANY', $this->split($url), $class);
+        $this->any($url, $class);
     }
 
     /**